Refactor main.rs — extract server startup logic into dedicated server.rs
- Created backend/src/server.rs containing:
  - Router builders (build_router, build_router_with_db)
  - HTTP handlers (health, root, metrics)
  - Middleware (build_cors, metrics_middleware)
  - Server run() entry point (DB pool, price cache, Redis,
    event bus, TCP bind, axum::serve)
- Stripped main.rs from 434 to 71 lines:
  - Kept: mod declarations, config loading, tracing/sentry init
  - Delegates to server::run(config)
- Re-exported build_router for backward compatibility with tests
- Fixed metrics_middleware signature for axum 0.7.9 API

Closes #994
